third_party/blink/web_tests/fast/dom/SelectorAPI/document-fragment-nth.html

<!DOCTYPE html>
<title>DocumentFragment.querySelector matching nth-child of fragment.</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
setup({ single_test: true });


var fragment = document.createDocumentFragment();
for (var i = 0; i < 12; i++) {
  var child = document.createElement("span");
  if (i == 11)
    child.id = "pass";
  fragment.appendChild(child);
}

assert_equals(fragment.querySelector(":nth-child(12)").id, "pass");
done();
</script>
